+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Cere Stats
 
 ## vNext
-...
+- [BE] Add superuser token allocation to faucet controller
 
 ## v0.28.0
 - [BE] Add ss58 prefix migration for block table